End of section 5

+++ src/AuctionService/Controllers/AuctionsController.cs
@@ -5,6 +5,7 @@ using AutoMapper;
 using AutoMapper.QueryableExtensions;
 using Contracts;
 using MassTransit;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 
@@ -52,13 +53,13 @@ public class AuctionsController : ControllerBase
         return _mapper.Map<AuctionDto>(auction);
 
     }
+    [Authorize]
     [HttpPost]
     public async Task<ActionResult<AuctionDto>> CreateAuction(CreateAuctionDto auctionDto)
     {
         var auction = _mapper.Map<Auction>(auctionDto);
 
-        //TODO add user
-        auction.Seller = "test";
+        auction.Seller = User.Identity.Name;
 
         _context.Auctions.Add(auction);
 
@@ -77,7 +78,7 @@ public class AuctionsController : ControllerBase
 
 
     }
-
+    [Authorize]
     [HttpPut("{id}")]
     public async Task<ActionResult> UpdateAuction(Guid id,UpdateAuctionDto updateAuctionDto)
     {
@@ -86,7 +87,8 @@ public class AuctionsController : ControllerBase
 
         if(auction == null) return NotFound();
 
-        //TODO check seller
+        if (auction.Seller != User.Identity.Name) return Forbid();
+
 
         auction.Item.Make =  updateAuctionDto.Make ?? auction.Item.Make;
         auction.Item.Model =  updateAuctionDto.Model ?? auction.Item.Model;
@@ -103,7 +105,7 @@ public class AuctionsController : ControllerBase
         return BadRequest("Problem saving changes");
 
     }
-
+    [Authorize]
     [HttpDelete("{id}")]
     public async Task<ActionResult> DeleteAuction(Guid id)
     {
@@ -111,7 +113,7 @@ public class AuctionsController : ControllerBase
 
       if(auction == null) return NotFound();
 
-      //TODO check seller
+      if (auction.Seller != User.Identity.Name) return Forbid();
 
 
        _context.Auctions.Remove(auction);
